Changed frame values to int
Hi! This is my first ever contribution to any code, so forgive me if it is wrong. In newer versions of blender the frame values for  setting custom frame range are expected to be int not float, so I thought it would be nice to fix this, as the current version of the addon is not working for this function.